Removed DbWeaponService.
