+++ src/server/app/zeldaplay/weapon/weapon.service.ts
@@ -1,22 +1,84 @@
 import { Injectable } from '@nestjs/common';
 import { Observable } from 'rxjs';
 
+import { DbService } from '@Db/db.service';
 import { DbWeapon } from '@DbModel/index';
-import { DbWeaponService } from './db-weapon/db-weapon.service';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class WeaponService {
-  constructor(private readonly dbService: DbWeaponService) {}
+  constructor(private readonly dbService: DbService) {}
 
   getWeapons(charId: string): Observable<DbWeapon[]> {
-    return this.dbService.getWeapons(charId);
+    return this.dbService.query<DbWeapon>(
+      `SELECT
+        id as "wId"
+        ,name as "wName"
+        ,modifier as "wModifier"
+        ,ammo as "wAmmo"
+        ,range as "wRange"
+        ,crit_damage as "wCritDamage"
+        ,crit_range as "wCritRange"
+        ,number_of_hits as "wNumberOfHits"
+        ,type as "wType"
+        ,damage as "wDamage"
+      FROM zeldaplay.weapons
+      WHERE character_id = $1`,
+      [charId]
+    );
   }
 
-  newWeapon(newWeap: DbWeapon, charId: string): Observable<DbWeapon> {
-    return this.dbService.newWeapon(newWeap, charId);
+  newWeapon(weapon: DbWeapon, charId: string): Observable<DbWeapon> {
+    return this.dbService
+      .query<DbWeapon>(
+        `INSERT INTO zeldaplay.weapons
+        (name, modifier, ammo, range, crit_damage, crit_range, number_of_hits, type, damage, character_id) VALUES
+        ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
+      RETURNING id as wId`,
+        [
+          weapon.wName,
+          weapon.wModifier,
+          weapon.wAmmo,
+          weapon.wRange,
+          weapon.wCritDamage,
+          weapon.wCritRange,
+          weapon.wNumberOfHits,
+          weapon.wType,
+          weapon.wDamage,
+          charId
+        ]
+      )
+      .pipe(map((weapons) => weapons[0]));
   }
 
-  updateWeapon(newWeap: DbWeapon): Observable<DbWeapon> {
-    return this.dbService.updateWeapon(newWeap);
+  updateWeapon(weapon: DbWeapon): Observable<DbWeapon> {
+    return this.dbService
+      .query<DbWeapon>(
+        `UPDATE zeldaplay.weapons as w
+        SET name = inWeap.name
+        ,modifier = inWeap.modifier
+        ,ammo = inWeap.ammo
+        ,range = inWeap.range
+        ,crit_damage = inWeap.crit_damage
+        ,crit_range = inWeap.crit_range
+        ,number_of_hits = inWeap.number_of_hits
+        ,type = inWeap.type
+        ,damage = inWeap.damage
+      FROM( VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9))
+      AS inWeap(name, modifier, ammo, range, crit_damage, crit_range, number_of_hits, type, damage)
+      WHERE w.id = ${weapon.wId}`,
+        [
+          weapon.wName,
+          weapon.wModifier,
+          weapon.wAmmo,
+          weapon.wRange,
+          weapon.wCritDamage,
+          weapon.wCritRange,
+          weapon.wNumberOfHits,
+          weapon.wType,
+          weapon.wDamage
+        ]
+      )
+      .pipe(map((weapons) => weapons[0]));
   }
 }
